Fix: Stored all the directories leading to the images inside a parent folder
Fix: Fixed a bug that could lead to a crash

+++ app/src/main/java/com/example/mangareader/Downloading/RemoveService.kt
@@ -53,12 +53,14 @@ class RemoveService : Service() {
                         break
                     }
                 }
-
                 if (targetChapter != null) {
                     val directory = File(targetChapter.imagesPath)
                     val files = directory.listFiles()
-                    for (file in files!!) {
-                        file.delete()
+                    if (files != null) {
+                        for (file in files) {
+                            if (file == null) continue
+                            file.delete()
+                        }
                     }
                 }
             }
